# api/cmd/api/openapi.json
{
  "components": {
    "schemas": {
      "CreateSandboxRequest": {
        "description": "Sandbox create request. CPU/memory may be overridden by the template's baked snapshot dimensions.",
        "properties": {
          "cpu": {
            "description": "Requested vCPUs",
            "type": "integer"
          },
          "disk_gb": {
            "description": "Optional rootfs size in GiB",
            "type": "integer"
          },
          "from_snapshot": {
            "description": "Restore from this snapshot ID instead of booting the template",
            "type": "string"
          },
          "memory_mb": {
            "description": "Requested memory in MiB",
            "type": "integer"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Free-form string metadata; keys starting with secret. are write-only",
            "type": "object"
          },
          "persistent": {
            "description": "true exempts the sandbox from the idle reaper",
            "type": "boolean"
          },
          "template": {
            "description": "Template name (e.g. base, code-interpreter, agent, browser). Public templates or ones owned by the caller's workspace only",
            "type": "string"
          },
          "ttl_seconds": {
            "description": "Idle TTL in seconds; null uses the platform default",
            "type": [
              "integer",
              "null"
            ]
          },
          "volumes": {
            "description": "Durable volumes to attach",
            "items": {
              "$ref": "#/components/schemas/VolumeMount"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "DatabaseCreateRequest": {
        "description": "Options for provisioning a managed database. All fields optional.",
        "properties": {
          "always_on": {
            "description": "Opt the database out of idle auto-suspend.",
            "type": "boolean"
          },
          "cpu": {
            "default": 2,
            "description": "vCPU count hint (the baked template snapshot governs the effective value).",
            "type": "integer"
          },
          "label": {
            "description": "Display label, surfaced in list/get.",
            "type": "string"
          },
          "memory_mb": {
            "default": 1024,
            "description": "Memory hint in MiB (the RAM tier/template governs the effective value).",
            "type": "integer"
          },
          "size": {
            "default": "1g",
            "description": "RAM tier. Create-time only — resizing later means clone with a new size.",
            "enum": [
              "1g",
              "4g",
              "16g"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
      "DatabaseInfo": {
        "description": "A managed PostgreSQL database (a postgres-16 family Firecracker sandbox). Connection fields are populated only while the database is running.",
        "properties": {
          "always_on": {
            "description": "true = opted out of idle auto-suspend.",
            "type": "boolean"
          },
          "broker_token": {
            "description": "Per-database REST query-broker bearer token (pds_pg_…); only returned while running.",
            "type": "string"
          },
          "broker_url": {
            "description": "Public REST query-broker base URL (/v1/databases/{id}/proxy).",
            "type": "string"
          },
          "cloned_from": {
            "description": "Source database id for clones/PITR restores.",
            "type": "string"
          },
          "branched_from": {
            "description": "Parent database id for reflink branches (TUSK T4.1).",
            "type": "string"
          },
          "connection_url": {
            "description": "Ready-to-use postgres:// URL (TLS required).",
            "type": "string"
          },
          "pooled_connection_url": {
            "description": "postgres:// URL routed through the guest's pgbouncer (transaction pooling) on port 6432. Recommended for serverless/agent workloads that open many short connections.",
            "type": "string"
          },
          "created_at": {
            "description": "Creation time, epoch seconds. Omitted when unknown.",
            "format": "int64",
            "type": "integer"
          },
          "database": {
            "description": "Default database name.",
            "type": "string"
          },
          "error": {
            "description": "Set when the VM is up but Postgres has not published credentials, or as a human-readable progress hint on async 202 responses.",
            "type": "string"
          },
          "failover_available": {
            "description": "Only on failed databases: whether a restore on another agent is possible.",
            "type": "boolean"
          },
          "failover_eta_seconds": {
            "description": "Expected restore time (RTO) in seconds when failover is available.",
            "type": "integer"
          },
          "failover_reason": {
            "description": "Why failover is or is not available.",
            "type": "string"
          },
          "host": {
            "description": "SNI host for native connections (<id>.db.pandastack.ai).",
            "type": "string"
          },
          "id": {
            "description": "Database id (currently doubles as the sandbox id).",
            "type": "string"
          },
          "label": {
            "description": "User-supplied display label.",
            "type": "string"
          },
          "password": {
            "description": "PostgreSQL password (rotates on restore/failover and on reset-credentials).",
            "type": "string"
          },
          "port": {
            "description": "PostgreSQL port (5432).",
            "type": "integer"
          },
          "sandbox_id": {
            "description": "Underlying sandbox id (explicit DB↔sandbox correlation).",
            "type": "string"
          },
          "size": {
            "description": "RAM tier, derived from the template.",
            "enum": [
              "1g",
              "4g",
              "16g"
            ],
            "type": "string"
          },
          "status": {
            "description": "Public status: provisioning | running | hibernated | failed | restoring (internal \"creating\" maps to provisioning).",
            "type": "string"
          },
          "template": {
            "description": "Backing template: postgres-16.",
            "type": "string"
          },
          "username": {
            "description": "PostgreSQL user.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "template"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "description": "Human-readable error message",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Sandbox": {
        "description": "A Firecracker microVM sandbox. secret.* metadata keys are redacted from all responses.",
        "properties": {
          "boot_mode": {
            "description": "How the sandbox booted",
            "enum": [
              "cold",
              "warm",
              "snapshot",
              null
            ],
            "type": [
              "string",
              "null"
            ]
          },
          "boot_ms": {
            "description": "Boot duration in milliseconds",
            "type": "integer"
          },
          "cpu": {
            "description": "vCPU count (may be overridden to match the template's baked snapshot)",
            "type": "integer"
          },
          "created_at": {
            "description": "Creation timestamp",
            "format": "date-time",
            "type": "string"
          },
          "from_snapshot": {
            "description": "Snapshot ID this sandbox was restored from, if any",
            "type": [
              "string",
              "null"
            ]
          },
          "guest_ip": {
            "description": "Guest IP inside its /30 subnet",
            "type": "string"
          },
          "host_tap": {
            "description": "Host-side TAP/netns identifier",
            "type": "string"
          },
          "id": {
            "description": "Sandbox UUID",
            "type": "string"
          },
          "mac": {
            "description": "Guest NIC MAC address",
            "type": "string"
          },
          "memory_mb": {
            "description": "Memory in MiB (may be overridden to match the template's baked snapshot)",
            "type": "integer"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Free-form string metadata; metadata.workspace is stamped server-side",
            "type": [
              "object",
              "null"
            ]
          },
          "status": {
            "description": "Current lifecycle status",
            "enum": [
              "creating",
              "running",
              "paused",
              "stopping",
              "deleted",
              "failed",
              "hibernated"
            ],
            "type": "string"
          },
          "template": {
            "description": "Template the sandbox was created from (e.g. base, code-interpreter)",
            "type": "string"
          },
          "vsock_cid": {
            "description": "vsock context ID for guest-agent communication",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "SuspendedError": {
        "properties": {
          "detail": {
            "description": "Human-readable explanation",
            "type": "string"
          },
          "error": {
            "description": "Short error code message",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "pds_... API token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "url": "https://pandastack.ai"
    },
    "description": "REST API for PandaStack — Firecracker microVM sandboxes, managed Postgres, git-driven apps, serverless functions, and outbound webhooks. Authenticate with a pds_ bearer token from the dashboard's API Tokens page.",
    "title": "PandaStack API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/databases": {
      "get": {
        "operationId": "listDatabases",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "count": {
                      "description": "Number of items.",
                      "type": "integer"
                    },
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/DatabaseInfo"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "items",
                    "count"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Databases in the workspace (connection fields are not included in list items)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Workspace not set."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Agent unreachable."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List all managed databases in the caller's workspace",
        "tags": [
          "Databases"
        ]
      },
      "post": {
        "operationId": "createDatabase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseCreateRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseInfo"
                }
              }
            },
            "description": "Database sandbox accepted. Poll GET /v1/databases/{id} until status is \"running\" (connection credentials appear then)."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid size (valid: 1g, 4g, 16g)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Workspace not set (unauthenticated)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Failed to decode the agent's sandbox response."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Agent error while creating the underlying sandbox."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Provision a new managed PostgreSQL 16 database (async; returns 202 with status \"provisioning\")",
        "tags": [
          "Databases"
        ]
      }
    },
...
    "/v1/sandboxes": {
      "get": {
        "operationId": "listSandboxes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Sandbox"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Array of sandboxes owned by the caller's workspace"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "List failed"
          }
        },
        "summary": "List the caller's sandboxes (workspace-scoped; internal pool/tombstone rows filtered out)",
        "tags": [
          "Sandboxes"
        ]
      },
      "post": {
        "operationId": "createSandbox",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sandbox"
                }
              }
            },
            "description": "Sandbox created and running"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed JSON body or invalid template name"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspendedError"
                }
              }
            },
            "description": "Account suspended, or template owned by another workspace"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Create failed"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No available compute node"
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Host memory admission refused (control plane places elsewhere)"
          }
        },
        "summary": "Create a sandbox (snapshot-restore boot, ~180ms p50)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
...
    "/v1/templates": {
      "get": {
        "operationId": "listTemplates",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Template"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Bare JSON array of templates (globals sorted first, then by name)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Catalog query failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List the template catalog (global first-party templates plus the caller's workspace-owned custom templates)",
        "tags": [
          "Templates"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://api.pandastack.ai"
    }
  ]
}